refactor: 重构图表导出功能为统一的组合式函数
- 删除旧的 utils/chart-export.ts 和 utils/export-excel.ts 工具文件
- 新增 useSendChartEmail 组合式函数，提供统一的图表导出API
- 更新所有图表组件使用新的导出方式
- 优化邮件发送服务，支持图表导出和邮件发送功能
- 修复图表导出相关的类型定义和模块引用问题
