+++ components/line-chart/index.vue
@@ -4,7 +4,7 @@
 </template>
 <script setup lang="ts">
 import { Chart } from '@antv/g2'
-import type { ExportChartOptions } from '~/utils/chart-export'
+
 defineOptions({
   name: 'LineChart'
 })
@@ -34,6 +34,9 @@ const emits = defineEmits(['renderChartStart', 'renderChartEnd'])
  */
 const chartInstance = ref<Chart | null>(null)
 
+// 图表导出功能
+const { exportChartAsBase64, downloadChartAsImage } = useSendChartEmail()
+
 // 暴露图表实例和导出方法给父组件
 defineExpose({
   chartInstance,
@@ -45,9 +48,7 @@ defineExpose({
     if (!chartInstance.value) {
       throw new Error('图表实例不存在')
     }
-
-    const { ChartExporter } = await import('~/utils/chart-export')
-    return ChartExporter.exportChartAsBase64(chartInstance.value as InstanceType<typeof Chart>, options)
+    return exportChartAsBase64(chartInstance.value as InstanceType<typeof Chart>, options)
   },
   /**
    * 下载图表
@@ -58,8 +59,7 @@ defineExpose({
     if (!chartInstance.value) {
       throw new Error('图表实例不存在')
     }
-    const { ChartExporter } = await import('~/utils/chart-export')
-    return ChartExporter.downloadChart(chartInstance.value as InstanceType<typeof Chart>, filename, options)
+    return downloadChartAsImage(chartInstance.value as InstanceType<typeof Chart>, filename, options)
   }
 })
