+++ server/api/sendEmail.post.ts
@@ -1,7 +1,7 @@
 import { EmailService } from '../service/emailService'
 
 export default defineEventHandler(async (event) => {
-  const body = await readBody<SendEmailDto.SendEmailOption>(event)
+  const body = await readBody<SendEmailDto.SendEmailOptions>(event)
   const { to, subject, html, text, cc, bcc, attachments } = body || {}
   const emailService = new EmailService()
   const messageId = await emailService.sendMail({
